+++ DependencyInjection/c975LPaymentExtension.php
@@ -35,6 +35,7 @@ class c975LPaymentExtension extends Extension
 
         $container->setParameter('c975_l_payment.live', $processedConfig['live']);
         $container->setParameter('c975_l_payment.site', $processedConfig['site']);
+        $container->setParameter('c975_l_payment.gdpr', $processedConfig['gdpr']);
         $container->setParameter('c975_l_payment.defaultCurrency', strtoupper($processedConfig['defaultCurrency']));
         $container->setParameter('c975_l_payment.vat', $processedConfig['vat'] * 100);
         $container->setParameter('c975_l_payment.stripeFeePercentage', $processedConfig['stripeFeePercentage']);